requirements.txt
absl-py==2.1.0
accelerate==0.27.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
ale-py==0.8.1
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
banks==2.1.2
bcrypt==5.0.0
beautifulsoup4==4.13.4
bigtree==0.31.0
bleach==6.0.0
blinker==1.9.0
bs4==0.0.1
cachetools==5.4.0
certifi==2025.4.26
cffi==1.15.1
charset-normalizer==3.4.2
clarabel==0.6.0
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.2
ctransformers==0.2.27
curl_cffi==0.12.0
cvxpy==1.4.1
cycler==0.12.1
dask==2025.9.1
dataclasses-json==0.6.7
datasets==4.0.0
DateTime==5.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.16
duckdb==1.2.2
ecdsa==0.19.2
ecos==2.0.12
efficientnet==1.0.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.1.2
executing==1.2.0
face_recognition_models==0.3.0
faiss-cpu==1.10.0
Faker==19.13.0
Farama-Notifications==0.0.4
fastapi==0.116.2
fastjsonschema==2.18.0
feast==0.53.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==24.3.25
fonttools==4.40.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2025.5.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-genai==1.32.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
griffe==1.7.2
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
gym==0.26.2
gym-anytrading==2.0.0
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h5py==3.11.0
helper==2.5.0
hf-xet==1.1.2
html5lib==1.1
httpcore==1.0.8
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
image-classifiers==1.0.0
imageio==2.34.2
importlib_metadata==8.7.0
importlib_resources==6.4.5
imutils==0.5.4
ipykernel==6.20.2
ipython==8.9.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.2.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.15
keras==3.11.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
llama-cloud==0.1.18
llama-cloud-services==0.6.12
llama-index==0.12.32
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.32
llama-index-embeddings-huggingface==0.5.3
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-legacy==0.9.48.post4
llama-index-llms-huggingface==0.5.0
llama-index-llms-huggingface-api==0.4.1
llama-index-llms-ollama==0.5.6
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-faiss==0.3.0
llama-parse==0.6.12
llamaindex-py-client==0.1.19
locket==1.0.0
lxml==4.9.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
mypy==1.18.2
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.35.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.0.2
notebook_shim==0.2.3
nsepy==0.8
numpy==2.2.6
oauthlib==3.2.2
ollama==0.4.8
openai==1.75.0
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.11.1
osqp==0.6.3
overrides==7.4.0
packaging==25.0
pandas==2.2.3
pandas-datareader==0.10.0
pandas_ta==0.3.14b0
pandasai==2.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.2
passlib==1.7.4
patchify==0.2.3
pathspec==0.12.1
patsy==0.5.6
pdfminer.six==20250327
pdfplumber==0.11.6
peewee==3.18.2
peft==0.17.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.3.7
plotly==5.16.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.1.0
Pygments==2.14.0
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
pyportfolioopt==1.5.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.24
python-pptx==1.0.2
pytz==2025.2
pywin32==305
pywinpty==2.0.11
PyYAML==6.0.2
pyzmq==25.0.0
qdldl==0.1.7.post0
referencing==0.30.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.15.3
scs==3.2.4.post1
seaborn==0.13.2
segmentation-models==1.0.0
Send2Trash==1.8.2
sentence-transformers==4.1.0
shtab==1.7.1
six==1.17.0
sklearn==0.0.post5
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.40
sqlglot==25.34.1
sqlglotrs==0.3.0
stable-baselines==2.10.2
stable_baselines3==2.4.0
stack-data==0.6.2
starlette==0.48.0
statsmodels==0.14.5
streamlit==1.44.1
striprtf==0.0.26
sympy==1.14.0
ta==0.11.0
tabulate==0.9.0
tavily-python==0.7.10
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.20.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2024.7.24
tiktoken==0.9.0
tinycss2==1.2.1
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
toolz==1.0.0
torch==2.7.0
torchvision==0.22.0
tornado==6.2
tqdm==4.67.1
traitlets==5.8.1
transformers==4.54.1
trl==0.9.6
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.15.0
tyro==0.8.6
tzdata==2025.2
ultralytics==8.3.141
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.0
uvicorn-worker==0.3.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.16.0
xgboost==3.0.3
XlsxWriter==3.1.2
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.65
zipp==3.23.0
zope.interface==6.1
zstandard==0.23.0
